Remove hard-coded headshot damage buff
The problem is that the "head" can be in a different position from the entity's pos, and hardcoding a headshot damage buff with the default MTG player model in mind is a bad idea.

The plan is to allow mods to use a yet-to-be-implemented on_hit callback to detect a headshot and deal extra damage to the target if deemed necessary.